+++ src/log_nondominating.c
@@ -34,20 +34,20 @@ static void lnd_evaluate_function(coco_problem_t *self, const double *x, double 
   coco_evaluate_function(coco_get_transform_inner_problem(self), x, y);
   data->number_of_evaluations++;
 
-  /* Open logfile if it is not alread open */
+  /* Open logfile if it is not already open */
   if (data->logfile == NULL) {
     data->logfile = fopen(data->path, "w");
     if (data->logfile == NULL) {
       char *buf;
       const char *error_format =
           "lnd_evaluate_function() failed to open log file '%s'.";
-      size_t buffer_size = snprintf(NULL, 0, error_format, data->path);
+      size_t buffer_size = (size_t)snprintf(NULL, 0, error_format, data->path);
       buf = (char *)coco_allocate_memory(buffer_size);
       snprintf(buf, buffer_size, error_format, data->path);
       coco_error(buf);
       coco_free_memory(buf); /* Never reached */
     }
-    fprintf(data->logfile, "# %zu variables  |  %zu objectives  |  func eval number\n",
+    fprintf(data->logfile, "# %ld variables  |  %ld objectives  |  func eval number\n",
             coco_get_number_of_variables(coco_get_transform_inner_problem(self)),
             coco_get_number_of_objectives(coco_get_transform_inner_problem(self)));
     
@@ -76,7 +76,7 @@ static void lnd_evaluate_function(coco_problem_t *self, const double *x, double 
           fprintf(data->logfile, "%13.10e\t", entry->var[j]);
       for (k=0; k < coco_get_number_of_objectives(coco_get_transform_inner_problem(self)); k++) /* all objective values of a solution */
           fprintf(data->logfile, "%13.10e\t", entry->obj[k]);
-      fprintf(data->logfile, "%zu", entry->birth);  /* its timestamp (FEval) */
+      fprintf(data->logfile, "%ld", entry->birth);  /* its timestamp (FEval) */
       fprintf(data->logfile, "\n");  /* go to the next line for another solution */
   }
   /********************************************************************************/
